Разбиение функций на модули
